User update

+++ src/main/java/com/testeWS/course/entities/Category.java
@@ -21,7 +21,7 @@ public class Category implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private  String name;
 	
 	@JsonIgnore
@@ -31,17 +31,17 @@ public class Category implements Serializable {
 	public Category() {
 	}
 	
-	public Category(Integer id, String name) {
+	public Category(Long id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
