+++ src/main/java/com/testeWS/course/entities/enums/OrderStatus.java
@@ -8,17 +8,17 @@ public enum OrderStatus {
 		DELIVERED(4),
 		CANCELED(5);
 	
-	private int code;
+	private long code;
 	
 	private OrderStatus(int code) {
 		this.code = code;
 	}
 	
-	public int getCode() {
+	public long getCode() {
 		return code;
 	}
 	
-	public static OrderStatus valueOf(int code) {
+	public static OrderStatus valueOf(long code) {
 		for(OrderStatus value : OrderStatus.values()) {
 			if(value.getCode() == code) {
 				return value;
